Update (index.html): Updated page text content, added image, added footer
The home page (index.html) was rather boring to engage with, so I added a cheeky photo of a cat in a chef's hat, updated the text to read better, and added a footer with website author and project information. I also included links to the final two recipe web pages (caprese panini and grilled vegetables).